@@ -1,6 +1,6 @@
 {
   "name": "cool-site",
-  "version": "69",
+  "version": "69.0.0",
   "description": "Internet",
   "main": "index.js",
   "scripts": {
@@ -10,10 +10,20 @@
   "author": "johnnybenson",
   "license": "MIT",
   "devDependencies": {
+    "@babel/core": "^7.5.0",
+    "@babel/preset-env": "^7.5.0",
+    "babel-loader": "^8.0.6",
     "copy-webpack-plugin": "^4.3.1",
-    "file-loader": "^0.8",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.10.0"
+    "css-loader": "^3.0.0",
+    "eslint": "^6.0.1",
+    "eslint-loader": "^2.2.1",
+    "file-loader": "^0.8.5",
+    "mini-css-extract-plugin": "^0.7.0",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^7.1.0",
+    "webpack": "^4.35.2",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "^3.7.2"
   },
   "dependencies": {
     "easy-scroll": "^0.5.1"
